server/requirements.txt
websockets>=12.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.53.0
python-dotenv>=1.0.0
google-adk>=0.1.6
torch
torchaudio
numpy
google-cloud-texttospeech
google-genai
websockets
asyncio